requirements.txt
asgiref==3.3.1
astor==0.8.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Django==3.1.7
gast==0.3.3
graphviz==0.16
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.8.0
joblib==1.0.1
kiwisolver==1.3.1
lmdb==1.1.1
matplotlib==3.2.1
netifaces==0.10.9
networkx==2.5
nltk==3.5
numpy==1.20.1
objgraph==3.5.0
opencv-python==4.2.0.32
paddlepaddle==2.0.0b0
pathlib==1.0.1
Pillow==8.1.2
prettytable==2.1.0
protobuf==3.15.6
pyclipper==1.2.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
rarfile==4.0
regex==2021.3.17
requests==2.25.1
scikit-image
scipy
Shapely==1.7.1
six==1.15.0
sqlparse==0.4.1
tifffile
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
zipp==3.4.1
